// src/lib/api/model/conditionType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Unplan API
 * Unplan Server API Documentation
 * OpenAPI spec version: v1.0.0
 */

export type ConditionType = (typeof ConditionType)[keyof typeof ConditionType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ConditionType = {
  DANGER: 'DANGER',
  LACK: 'LACK',
  OPTIMAL: 'OPTIMAL',
  EXCESS: 'EXCESS',
} as const;