// src/lib/api/model/onboardingRequestTransportationsItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Unplan API
 * Unplan Server API Documentation
 * OpenAPI spec version: v1.0.0
 */

export type OnboardingRequestTransportationsItem =
  (typeof OnboardingRequestTransportationsItem)[keyof typeof OnboardingRequestTransportationsItem];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OnboardingRequestTransportationsItem = {
  WALK: 'WALK',
  BICYCLE: 'BICYCLE',
  PUBLIC_TRANSPORT: 'PUBLIC_TRANSPORT',
  CAR: 'CAR',
} as const;