// src/lib/api/model/updateMethodsDefaultMethodsItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Unplan API
 * Unplan Server API Documentation
 * OpenAPI spec version: v1.0.0
 */

export type UpdateMethodsDefaultMethodsItem =
  (typeof UpdateMethodsDefaultMethodsItem)[keyof typeof UpdateMethodsDefaultMethodsItem];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UpdateMethodsDefaultMethodsItem = {
  NAP: 'NAP',
  MUSIC: 'MUSIC',
  WALK: 'WALK',
  STRETCHING: 'STRETCHING',
  FOOD: 'FOOD',
} as const;